requirements.txt
openai
agentscope
fastapi
uvicorn
python-dotenv
pydantic